requirements.txt
asgiref==3.8.1
beautifulsoup4==4.11.1
certifi==2022.9.24
charset-normalizer==2.1.1
Django==5.1.6
django-bootstrap4==22.3
django-filter==23.1
django-widget-tweaks==1.4.12
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
docopt==0.6.2
idna==3.4
mysqlclient==2.1.1
pipreqs==0.4.11
PyJWT==2.9.0
requests==2.28.1
soupsieve==2.3.2.post1
sqlparse==0.4.3
typing_extensions==4.12.2
urllib3==1.26.13
yarg==0.1.9
